requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
elevenlabs==2.3.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.18.0
fsspec==2025.5.1
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
instagrapi==1.19.0
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai-whisper==20240930
pillow==11.2.1
praw==7.8.1
prawcore==2.4.0
pydantic==2.11.7
pydantic_core==2.33.2
regex==2024.11.6
requests==2.32.4
sniffio==1.3.1
sympy==1.14.0
tiktoken==0.9.0
toml==0.10.2
torch==2.7.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
update-checker==0.18.0
urllib3==2.4.0
websocket-client==1.8.0
websockets==15.0.1
yt-dlp==2025.6.9
